refactor and remove stuff
